+++ apps/dashboard/src/app/resources/edit/edit-resource.service.ts
@@ -1,10 +1,38 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject, catchError, forkJoin, map, Observable, of, switchMap, take, tap } from 'rxjs';
-import { Classification, deDuplicateList, Resource, UserClassification } from '@nuclia/core';
+import {
+  BehaviorSubject,
+  catchError,
+  filter,
+  forkJoin,
+  map,
+  Observable,
+  of,
+  switchMap,
+  take,
+  tap,
+  throwError,
+} from 'rxjs';
+import {
+  Classification,
+  deDuplicateList,
+  FIELD_TYPE,
+  FieldId,
+  FileFieldData,
+  IFieldData,
+  KeywordSetField,
+  LinkField,
+  Resource,
+  ResourceData,
+  ResourceField,
+  TextField,
+  UserClassification,
+} from '@nuclia/core';
 import { SDKService } from '@flaps/core';
-import { SisToastService } from '@nuclia/sistema';
+import { SisModalService, SisToastService } from '@nuclia/sistema';
+import { TranslateService } from '@ngx-translate/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
-export type EditResourceView = 'profile' | 'classification';
+export type EditResourceView = 'profile' | 'classification' | 'add-field';
 
 @Injectable({
   providedIn: 'root',
@@ -12,22 +40,54 @@ export type EditResourceView = 'profile' | 'classification';
 export class EditResourceService {
   private _resource = new BehaviorSubject<Resource | null>(null);
   private _currentView = new BehaviorSubject<EditResourceView | null>(null);
+  private _currentField = new BehaviorSubject<FieldId | 'profile'>('profile');
 
-  resource = this._resource.asObservable();
-  currentView = this._currentView.asObservable();
+  currentView: Observable<EditResourceView | null> = this._currentView.asObservable();
+  currentField: Observable<FieldId | 'profile'> = this._currentField.asObservable();
+  resource: Observable<Resource | null> = this._resource.asObservable();
+  fields: Observable<ResourceField[]> = this.resource.pipe(
+    map((resource) =>
+      Object.entries(resource?.data || {}).reduce((list, [type, dict]) => {
+        return list.concat(
+          Object.entries(dict).map(([fieldId, field]) => ({
+            ...field,
+            field_id: fieldId,
+            field_type: type.slice(0, -1),
+          })),
+        );
+      }, [] as ResourceField[]),
+    ),
+  );
 
-  constructor(private sdk: SDKService, private toaster: SisToastService) {}
+  constructor(
+    private sdk: SDKService,
+    private router: Router,
+    private toaster: SisToastService,
+    private modalService: SisModalService,
+    private translate: TranslateService,
+  ) {}
 
-  loadResource(resourceId: string) {
-    this.sdk.currentKb
-      .pipe(
-        take(1),
-        switchMap((kb) => kb.getResource(resourceId)),
-      )
-      .subscribe((resource) => this._resource.next(resource));
+  private _loadResource(resourceId: string): Observable<Resource> {
+    return this.sdk.currentKb.pipe(
+      take(1),
+      switchMap((kb) => kb.getResource(resourceId)),
+      tap((resource) => this._resource.next(resource)),
+    );
   }
 
-  save(partialResource: Partial<Resource>): Observable<void | null> {
+  loadResource(resourceId: string) {
+    this._loadResource(resourceId).subscribe();
+  }
+
+  getField(fieldType: keyof ResourceData, fieldId: string): Observable<IFieldData> {
+    return this.resource.pipe(
+      filter((resource) => !!resource),
+      map((resource) => resource as Resource),
+      map((resource) => resource.data[fieldType]?.[fieldId] || {}),
+    );
+  }
+
+  savePartialResource(partialResource: Partial<Resource>): Observable<void | null> {
     const currentResource = this._resource.value;
     if (!currentResource) {
       return of(null);
@@ -40,9 +100,8 @@ export class EditResourceService {
       ),
     ]).pipe(
       catchError((error) => {
-        this.toaster.error(error);
-        this._resource.next(currentResource);
-        return of(null);
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
       }),
       map(() => this.toaster.success('resource.save-successful')),
     );
@@ -52,9 +111,14 @@ export class EditResourceService {
     this._currentView.next(view);
   }
 
+  setCurrentField(field: FieldId | 'profile') {
+    this._currentField.next(field);
+  }
+
   reset() {
     this._resource.next(null);
     this._currentView.next(null);
+    this._currentField.next('profile');
   }
 
   getClassificationsPayload(labels: Classification[]): UserClassification[] {
@@ -73,5 +137,223 @@ export class EditResourceService {
       .filter((label) => !labels.some((l) => l.labelset === label.labelset && l.label === label.label))
       .map((label) => ({ ...label, cancelled_by_user: true }));
     return [...userClassifications, ...cancellations];
+  }
+
+  addField(
+    fieldType: FIELD_TYPE,
+    fieldId: string,
+    data: TextField | LinkField | KeywordSetField,
+  ): Observable<void | null> {
+    const currentResource = this._resource.value;
+    if (!currentResource) {
+      return of(null);
+    }
+
+    const dataKey = this.getDataKeyFromFieldType(fieldType);
+    const updatedData: ResourceData = dataKey
+      ? {
+          ...currentResource.data,
+          [dataKey]: {
+            ...currentResource.data[dataKey],
+            [fieldId]: {
+              value: data,
+            },
+          },
+        }
+      : currentResource.data;
+    return forkJoin([
+      currentResource.addField(fieldType, fieldId, data),
+      this.sdk.currentKb.pipe(
+        take(1),
+        tap((kb) => this._resource.next(kb.getResourceFromData({ ...currentResource, data: updatedData }))),
+      ),
+    ]).pipe(
+      catchError((error) => {
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
+      }),
+      map(() => this.toaster.success('resource.field.addition-successful')),
+    );
+  }
+
+  addFile(fieldId: string, file: File): Observable<void | null> {
+    const currentResource = this._resource.value;
+    if (!currentResource) {
+      return of(null);
+    }
+
+    const dataKey = this.getDataKeyFromFieldType(FIELD_TYPE.file);
+    const updatedData: ResourceData = dataKey
+      ? {
+          ...currentResource.data,
+          [dataKey]: {
+            ...currentResource.data[dataKey],
+            [fieldId]: this.getFileFieldData(file),
+          },
+        }
+      : currentResource.data;
+    return currentResource.upload(fieldId, file).pipe(
+      switchMap(() => this.sdk.currentKb.pipe(take(1))),
+      tap((kb) => this._resource.next(kb.getResourceFromData({ ...currentResource, data: updatedData }))),
+      catchError((error) => {
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
+      }),
+      map(() => this.toaster.success('resource.field.update-successful')),
+    );
+  }
+
+  updateField(
+    fieldType: FIELD_TYPE,
+    fieldId: string,
+    data: TextField | LinkField | KeywordSetField,
+  ): Observable<void | null> {
+    const currentResource = this._resource.value;
+    if (!currentResource) {
+      return of(null);
+    }
+
+    const updatedData: ResourceData = this.getUpdatedData(fieldType, currentResource.data, (fields, [id, field]) => {
+      if (id !== fieldId) {
+        fields[id] = field;
+      } else {
+        fields[id] = {
+          value: data,
+        };
+      }
+      return fields;
+    });
+    return forkJoin([
+      currentResource.updateField(fieldType, fieldId, data),
+      this.sdk.currentKb.pipe(
+        take(1),
+        tap((kb) => this._resource.next(kb.getResourceFromData({ ...currentResource, data: updatedData }))),
+      ),
+    ]).pipe(
+      catchError((error) => {
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
+      }),
+      map(() => this.toaster.success('resource.field.update-successful')),
+    );
+  }
+
+  updateFile(fieldId: string, file: File): Observable<void | null> {
+    const currentResource = this._resource.value;
+    if (!currentResource) {
+      return of(null);
+    }
+
+    const updatedData: ResourceData = this.getUpdatedData(
+      FIELD_TYPE.file,
+      currentResource.data,
+      (fields, [id, field]) => {
+        if (id !== fieldId) {
+          fields[id] = field;
+        } else {
+          fields[id] = this.getFileFieldData(file);
+        }
+        return fields;
+      },
+    );
+    return currentResource.deleteField(FIELD_TYPE.file, fieldId).pipe(
+      switchMap(() => currentResource.upload(fieldId, file)),
+      switchMap(() => this.sdk.currentKb.pipe(take(1))),
+      tap((kb) => this._resource.next(kb.getResourceFromData({ ...currentResource, data: updatedData }))),
+      catchError((error) => {
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
+      }),
+      map(() => this.toaster.success('resource.field.update-successful')),
+    );
+  }
+
+  confirmAndDelete(fieldType: FIELD_TYPE, fieldId: string, route: ActivatedRoute): Observable<void | null> {
+    return this.modalService
+      .openConfirm({
+        title: this.translate.instant('resource.field.delete-confirm-title', {
+          type: this.translate.instant(`resource.field-${fieldType}`),
+        }),
+        confirmLabel: 'generic.delete',
+        description: this.translate.instant('resource.field.delete-confirm-description', {
+          type: this.translate.instant(`resource.field-${fieldType}`),
+        }),
+        isDestructive: true,
+      })
+      .onClose.pipe(
+        filter((confirm) => !!confirm),
+        switchMap(() => this.deleteField(fieldType, fieldId)),
+        tap((done) => {
+          if (done !== null) {
+            this.router.navigate(['../../profile'], { relativeTo: route });
+          }
+        }),
+      );
+  }
+
+  private deleteField(fieldType: FIELD_TYPE, fieldId: string): Observable<void | null> {
+    const currentResource = this._resource.value;
+    if (!currentResource) {
+      return of(null);
+    }
+
+    const updatedData: ResourceData = this.getUpdatedData(fieldType, currentResource.data, (fields, [id, field]) => {
+      if (id !== fieldId) {
+        fields[id] = field;
+      }
+      return fields;
+    });
+    return forkJoin([
+      currentResource.deleteField(fieldType, fieldId),
+      this.sdk.currentKb.pipe(
+        take(1),
+        tap((kb) => this._resource.next(kb.getResourceFromData({ ...currentResource, data: updatedData }))),
+      ),
+    ]).pipe(
+      catchError((error) => {
+        this.toaster.error('generic.error.oops');
+        return throwError(() => error);
+      }),
+      map(() => this.toaster.success('resource.field.delete-successful')),
+    );
+  }
+
+  getDataKeyFromFieldType(fieldType: FIELD_TYPE): keyof ResourceData | null {
+    // Currently in our models, there are more FIELD_TYPEs than ResourceData keys, so we need the switch for typing reason
+    switch (fieldType) {
+      case FIELD_TYPE.text:
+      case FIELD_TYPE.file:
+      case FIELD_TYPE.link:
+      case FIELD_TYPE.keywordset:
+        return `${fieldType}s`;
+      default:
+        return null;
+    }
+  }
+
+  private getUpdatedData(
+    fieldType: FIELD_TYPE,
+    currentData: ResourceData,
+    reduceCallback: (previous: any, current: [string, any]) => ResourceData,
+  ): ResourceData {
+    const dataKey = this.getDataKeyFromFieldType(fieldType);
+    return dataKey
+      ? {
+          ...currentData,
+          [dataKey]: Object.entries(currentData[dataKey] || {}).reduce(reduceCallback, {} as any),
+        }
+      : currentData;
+  }
+
+  private getFileFieldData(file: File): FileFieldData {
+    return {
+      value: {
+        file: {
+          filename: file.name,
+          content_type: file.type,
+          size: file.size,
+        },
+      },
+    };
   }
 }