+++ apps/dashboard/src/app/resources/edit/file/file.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-resource-file',
-  templateUrl: 'file.component.html',
-})
-export class ResourceFileComponent {}