+++ apps/dashboard/src/app/resources/edit/link/link.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-resource-link',
-  templateUrl: 'link.component.html',
-})
-export class ResourceLinkComponent {}