+++ apps/dashboard/src/app/resources/edit/text/text.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-resource-text',
-  templateUrl: 'text.component.html',
-})
-export class ResourceTextComponent {}