+++ apps/dashboard/src/app/resources/resources.module.ts
@@ -20,9 +20,9 @@ import { TableHeaderComponent } from './table-header/table-header.component';
 import { EditResourceComponent } from './edit/edit-resource.component';
 import { ResourcesComponent } from './resources.component';
 import { ResourceProfileComponent } from './edit/profile/profile.component';
-import { ResourceTextComponent } from './edit/text/text.component';
-import { ResourceLinkComponent } from './edit/link/link.component';
-import { ResourceFileComponent } from './edit/file/file.component';
+import { ResourceTextComponent } from './edit/profile/text/text.component';
+import { ResourceLinkComponent } from './edit/profile/link/link.component';
+import { ResourceFileComponent } from './edit/profile/file/file.component';
 import { LabelModule } from '../components/label/label.module';
 import { HintModule } from '../components/hint/hint.module';
 import {
@@ -47,6 +47,8 @@ import { DatasetImportComponent } from './sample-dataset/dataset-import.componen
 import { LoadingModalComponent } from './sample-dataset/loading-modal/loading-modal.component';
 import { UploadButtonComponent } from './upload-button/upload-button.component';
 import { ResourceClassificationComponent } from './edit/classification/resource-classification.component';
+import { AddFieldComponent } from './edit/add-field/add-field.component';
+import { DropzoneComponent } from './dropzone/dropzone.component';
 
 const Components = [
   ResourceListComponent,
@@ -84,24 +86,26 @@ const ROUTES: Routes = [
           {
             path: 'profile',
             component: ResourceProfileComponent,
-            children: [
-              {
-                path: 'text',
-                component: ResourceTextComponent,
-              },
-              {
-                path: 'link',
-                component: ResourceLinkComponent,
-              },
-              {
-                path: 'file',
-                component: ResourceFileComponent,
-              },
-            ],
           },
           {
             path: 'classification',
             component: ResourceClassificationComponent,
+          },
+          {
+            path: 'add-field',
+            component: AddFieldComponent,
+          },
+          {
+            path: 'text/:fieldId',
+            component: ResourceTextComponent,
+          },
+          {
+            path: 'link/:fieldId',
+            component: ResourceLinkComponent,
+          },
+          {
+            path: 'file/:fieldId',
+            component: ResourceFileComponent,
           },
         ],
       },
@@ -153,6 +157,8 @@ const ROUTES: Routes = [
     LoadingModalComponent,
     UploadButtonComponent,
     ResourceClassificationComponent,
+    AddFieldComponent,
+    DropzoneComponent,
   ],
   exports: [],
 })