+++ libs/sdk-core/src/lib/db/resource/resource.models.ts
@@ -290,6 +290,8 @@ export interface Sentence {
 }
 
 export interface CloudLink {
+  filename?: string;
+  md5?: string;
   uri?: string;
   size?: number;
   content_type?: string;